requirements.txt
lxml
json-rpc
pymongo
pika
werkzeug
requests
redis
sklearn
python-dateutil
newspaper3k